+++ RayTracerV3/lightFactory.h
@@ -9,6 +9,6 @@ class LightFactory {
 public:
     static AbstractLight* createPointLight(const Vec3& position, const Color& color);
     static AbstractLight* createDirectionalLight(const Vec3& direction, const Color& color);
-    static AbstractLight* createAttributePointLight(const Vec3& position, const Color& color, float c1, float c2, float c3);
-    static AbstractLight* createAttributeDirectionalLight(const Vec3& direction, const Color& color, float c1, float c2, float c3);
+    static AbstractLight* createAttributePointLight(const Vec3& position, const Color& color, double c1, double c2, double c3);
+    static AbstractLight* createAttributeDirectionalLight(const Vec3& direction, const Color& color, double c1, double c2, double c3);
 };
