+++ RayTracerV3/materialFactory.h
@@ -6,5 +6,5 @@
 class MaterialFactory {
 public:
     static IMaterial* createRGBMaterial(const Color& color);
-    static IMaterial* createPhongMaterial(const Color& od, const Color& os, float ka = 0, float kd = 0, float ks = 0, float n = 0);
+    static IMaterial* createPhongMaterial(const Color& od, const Color& os, double ka = 0, double kd = 0, double ks = 0, double n = 0);
 };
